Switched to boost::result_of and added an overloaded const operator.
decltype support by compilers is still unperfect and at some point the recursive complexity is so high that type deduction cannot be properly done. Fortunately, Boost.Utility provides us with a result_of template class that helps a lot in this aspect. This modification doesn't change the way the library is used.